ld: add the hooks for the pointer scaling
This changes how we handle word addressed systems. We remove all the
automatic assumptions about scaling and instead expect the object file
to describe scaling. This is necessary to cope with byte pointers.

As an added bonus we in theory now have the linker bits needed to support
BCPL and B pointers on byte addressed machines.