pdp4: start implementing the bits for PDP 18bit mode
This will need a bunch of core changes which is kind of the point. This isn't
likely to be a very useful PDP4/7/9/15 assembler but it lets us start to sort
out link support for bank/page style addressing, bigger than 16bit words, and
some other stuff like constant merging we'll need to get right for some of
the odder processors.