6502: introduce a separate parsing function for constants/values
There are a load of cases where the fact "1234" is implicitly an address
are handled by constification on most targets but the 6502 has other
ambiguities when comma is involved because of the comma in addressing forms.

So that we can reliably allow stuff like

	.byte 12,13

not just
	.byte #12,#13

introduce a second simple parser for those cases where register addressing
forms are not valid.

Closes: #9